pyproject.toml
[project]
name = "cyber-lighthouse"
version = "0.1.0"
description = "Real-time threat intelligence monitoring system with AI analysis"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "feedparser>=6.0.12",
    "google-genai>=1.69.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]
